+++ packages/stacked_generator/lib/src/generators/extensions/list_utils_extension.dart
@@ -1,0 +1,10 @@
+extension LoggerClassGeneratorExtension on List<String> {
+  String get addCheckForReleaseModeToEachLogger {
+    final _multiLoggers = StringBuffer();
+
+    this.forEach((element) {
+      _multiLoggers.write(" if(kReleaseMode) $element(),");
+    });
+    return _multiLoggers.toString();
+  }
+}
